requirements.txt
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
Pillow==7.2.0
requests==2.24.0

pip~=20.2.2
channels~=2.4.0